Use byte for version

+++ stack-lang/pickle/Format.scala
@@ -5,11 +5,15 @@ object Format:
   // Format: MAJOR.MINOR
   // - Increment MAJOR for breaking changes (incompatible format)
   // - Increment MINOR for backward-compatible changes
-  final val MAJOR_VERSION: Int = 1
-  final val MINOR_VERSION: Int = 0
+  final val MAJOR_VERSION: Byte = 1
+  final val MINOR_VERSION: Byte = 0
 
   // Magic number to identify SAST files (ASCII: "SAST")
   final val MAGIC_NUMBER: Int = 0x53415354
+
+  //----------------------------------------------------------------------------
+  // Format Tags
+  //----------------------------------------------------------------------------
 
   final val Literal     : Byte = 1
   final val RecordLit   : Byte = 2
